Coding/requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3
imutils==0.5.4
pillow==10.0.0
requests==2.31.0
https://github.com/UB-Mannheim/tesseract/wiki